еще раз поправил навигацию, поправил фильтры, теперь диалоги показываются при каждом лайке

+++ feature/home/src/main/java/com/redpine/home/domain/utils/Filters.kt
@@ -1,11 +1,9 @@
 package com.redpine.home.domain.utils
 
-import javax.inject.Singleton
-
 data class Filters(
     val isMale: Boolean?,
-    val minAge: Long,
-    val maxAge: Long,
+    val minAge: Int,
+    val maxAge: Int,
     val size: List<String>,
     val color: List<String>,
     val character: List<String>,
